+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 @Data
 public class User {
-    private int id;
+    private Long id;
 
     @NotBlank
     @Email
@@ -26,5 +26,6 @@ public class User {
     @PastOrPresent
     private LocalDate birthday;
 
-    private final Set<Integer> friends = new HashSet<>();
+    private final Set<Long> friends = new HashSet<>();
+
 }
